refactor: directories to flat
