+++ server/internal/handler/health.go
